Refactor extras.models
