+++ netbox/extras/utils.py
@@ -22,6 +22,22 @@ def is_taggable(obj):
     return False
 
 
+def image_upload(instance, filename):
+    """
+    Return a path for uploading image attchments.
+    """
+    path = 'image-attachments/'
+
+    # Rename the file to the provided name, if any. Attempt to preserve the file extension.
+    extension = filename.rsplit('.')[-1].lower()
+    if instance.name and extension in ['bmp', 'gif', 'jpeg', 'jpg', 'png']:
+        filename = '.'.join([instance.name, extension])
+    elif instance.name:
+        filename = instance.name
+
+    return '{}{}_{}_{}'.format(path, instance.content_type.name, instance.object_id, filename)
+
+
 @deconstructible
 class FeatureQuery:
     """
